requirements.txt
# Django Core
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.5.0

# Computer Vision and Image Processing
opencv-python==4.10.0.84
mediapipe==0.10.20
numpy==1.26.4
Pillow==11.0.0
scikit-image==0.24.0

# Machine Learning (optional for future enhancements)
tensorflow==2.18.0

# Database
psycopg2-binary==2.9.10

# Configuration Management
python-decouple==3.8

# Excel Export
openpyxl==3.1.5
pandas==2.2.3

# Development Tools (optional)
pytest==8.3.4
pytest-django==4.9.0
